backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
neo4j==5.27.0
pydantic==2.10.4
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
google-generativeai==0.8.4
